### 2026-03-23 #10249 @editorialbot
[PRE REVIEW]: Chemiscope 1.0: interactive exploration of atomistic data from analysis to dissemination
**Submitting author:** @sofiia-chorna (<a href="http://orcid.org/0009-0008-7426-0856">Sofiia Chorna</a>)
**Repository:** https://github.com/lab-cosmo/chemiscope
**Branch with paper.md** (empty if default branch): paper-1.0
**Version:** v1.0.0
