### 2026-03-25 #10249 @Kevin-Mattheus-Moerman
Note to editors, this is the second submission for the Chemiscope project. This is first paper from 2020: https://joss.theoj.org/papers/10.21105/joss.02117. The authors have said that the release for which publication is sought here has added:

> "...adding cross-platform support, performance optimizations for large datasets, and direct compatibility with popular atomistic Python libraries like ASE, MDAnalysis and Chemfiles"
