### 2026-03-30 f181362
Add maintainers in README (#552)
Feels a bit weird to be thanking myself for maintaining the package, but
this is the blurb we decided on ^_^.
