+++ README.md
@@ -135,3 +135,7 @@ If you are interested in contributing to chemiscope, please have a look at our
 Chemiscope itself is distributed under the 3-Clauses BSD license. By
 contributing to this repository, you agree to distribute your contributions
 under the same license.
+
+This project is [maintained](https://github.com/lab-cosmo/.github/blob/main/Maintainers.md) by @ceriottm and @sofiia-chorna, who will reply to issues and pull requests opened on this repository as soon as possible. You can mention them directly if you did not receive an answer after a couple of days.
+
+It was previously maintained by @Luthaf (2019 to 2026), many thanks to them for their work making chemiscope awesome!
