### 2026-04-06 #10249 @abhishektiwari
👋 @braniii @ClementCaporal @wladerer, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Chemiscope 1.0: interactive exploration of atomistic data from analysis to dissemination".

Thanks
Abhi
