### 2026-04-09 #10380 @rvhonorato
> Contribution and authorship: Has the submitting author (@sofiia-chorna) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Looking at the [contribution graph](https://github.com/lab-cosmo/chemiscope/graphs/contributors) I see that the author is not the first on the list. I understand very well that the number of commits is not equivalent to valuable contributions, one line can change how a whole project works - but considering the big difference in number from the first one (560) to yours (53) could you please comment on it @sofiia-chorna?
