### 2026-04-09 #10380 @sofiia-chorna
Hello @rvhonorato! Thank you for your comment. The top contributor is the original creator of Chemiscope, and the bulk of those 560 commits correspond to the development covered by the [original 2020 JOSS publication](https://doi.org/10.21105/joss.02117). This paper is specifically about the new features introduced in the 1.0 release, in which I was involved either as the implementer or as a code reviewer. You can verify this through my [authored PRs](https://github.com/lab-cosmo/chemiscope/pulls?q=is%3Apr+is%3Aclosed+author%3Asofiia-chorna) and [reviewed PRs](https://github.com/lab-cosmo/chemiscope/pulls?q=is%3Apr+is%3Aclosed+reviewed-by%3Asofiia-chorna), which together cover the new functionality discussed in the paper.
