### 2026-04-09 #10380 @Luthaf
Initial author of chemiscope here, most of the recent work on the project was indeed done and coordinated by @sofiia-chorna!
