### 2026-04-11 #10380 @ceriottm
"Paying the bills" author here, and also decently active amateur contributor. I also stand firmly behind the fact that @sofiia-chorna earned the first spot in this 1.0 release, designing new features, doing lots of grunt work and overall becoming the de facto lead maintainer.
