### 2026-04-20 f51d7d5
fix colorbar crash with many symbol categories
