### 2026-04-20 2b099c4
fix duplicate legend items bug
