### 2026-04-20 4ce7aaf
Fixes of colorbar crash and duplicated legend items (#555)
fix two bugs:
1.
https://github.com/lab-cosmo/chemiscope/commit/f51d7d53e74a3abc8bdc87ee19d0ccca8cd7fb3d
- crash when colorbar.len went negative with many symbol categories,
plus fix padding if there is no legend
2.
https://github.com/lab-cosmo/chemiscope/commit/2b099c42d1e975cc274046cb35ff80a75fab8157
- duplicate legend items bug
