+++ docs/_downloads/cabbaa49808f095fba403265492cad54/5-shapes.py
@@ -1,0 +1,401 @@
+"""
+Shape visualization
+===================
+
+This example demonstrates how to visualize structure and atomic
+properties in the structure panel, using different types of
+predefined shapes (ellipsoids for tensors, arrows for vectors).
+The example also shows how to define custom shapes.
+
+Note that the same parameters can be used with :py:func:`chemiscope.show` to visualize
+an interactive widget in a Jupyter notebook.
+"""
+
+# %%
+#
+
+import ase.io
+import numpy as np
+from scipy.spatial.transform import Rotation
+
+import chemiscope
+
+
+# %%
+#
+# Loads a dataset of structures
+
+structures = ase.io.read("data/alpha-mu.xyz", ":")
+
+quaternions = []
+# converts the arrays from the format they are stored in to an array
+# format that can be processed by the ASE utilities
+for structure in structures:
+    structure.positions += structure.cell.diagonal() * 0.5
+    structure.arrays["alpha"] = np.array(
+        [
+            [axx, ayy, azz, axy, axz, ayz]
+            for (axx, ayy, azz, axy, axz, ayz) in zip(
+                structure.arrays["axx"],
+                structure.arrays["ayy"],
+                structure.arrays["azz"],
+                structure.arrays["axy"],
+                structure.arrays["axz"],
+                structure.arrays["ayz"],
+                strict=True,
+            )
+        ]
+    )
+
+    # interatomic separations (used to orient "stuff" later on)
+    dists = structure.get_all_distances(mic=True)
+    np.fill_diagonal(dists, np.max(dists))
+    for i in range(len(structure)):
+        nneigh = dists[i].argmin()
+        vec = structure.get_distance(i, nneigh, vector=True, mic=True)
+        vec /= np.linalg.norm(vec)
+        quat = Rotation.align_vectors([np.array([0, 0, 1])], [vec])[0].as_quat()
+        quaternions.append(quat)
+
+# %%
+#
+# Shapes generation
+# -----------------
+#
+# Here we define shapes that will later be used to create the input.
+# Input generation can also be achieved as a single call, but in practice
+# it is wise to define separate entities for better readability.
+
+# %%
+#
+# Cubes with smooth shading, centered on atoms. These are created as
+# "custom" shapes and then are just scaled to atom-dependent sizes:
+
+atom_sizes = {"O": 0.4, "N": 0.5, "C": 0.45, "H": 0.2}
+smooth_cubes = dict(
+    kind="custom",
+    parameters={
+        "global": {
+            # this list defines the vertices
+            "vertices": [
+                [-1, -1, -1],
+                [1, -1, -1],
+                [-1, 1, -1],
+                [1, 1, -1],
+                [-1, -1, 1],
+                [1, -1, 1],
+                [-1, 1, 1],
+                [1, 1, 1],
+            ],
+            # and these are the indices of the vertices that form the triangular mesh
+            "simplices": [
+                [0, 2, 1],
+                [1, 2, 3],
+                [4, 6, 5],
+                [5, 6, 7],
+                [0, 1, 4],
+                [0, 4, 2],
+                [1, 3, 5],
+                [2, 6, 3],
+                [1, 5, 4],
+                [2, 4, 6],
+                [3, 7, 5],
+                [3, 6, 7],
+            ],
+        },
+        # the cube is used at each atomic position, only difference being the scaling
+        "atom": [
+            {"scale": atom_sizes[label]}
+            for label in (list(structures[0].symbols) + list(structures[1].symbols))
+        ],
+    },
+)
+
+# %%
+#
+# Structure-based shapes. Also demonstrates how to achieve sharp-edge shading.
+# It requires defining multiple times the same vertices:
+
[…]
