+++ docs/_sources/examples/9-stk-custom-bonds.rst.txt
@@ -1,0 +1,594 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "examples/9-stk-custom-bonds.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_examples_9-stk-custom-bonds.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_examples_9-stk-custom-bonds.py:
+
+
+Using stk structures and showing custom bonds
+=============================================
+
+This example demonstrates interfacing `stk <https://stk.readthedocs.io/en/stable/>`_
+with chemiscope and the use of shapes to show custom bonding in a molecule.
+
+`stk <https://stk.readthedocs.io/en/stable/>`_ comes with topology/bonding
+information by default (using the cheminformatic software rdkit).
+
+We use `stko <https://stko-docs.readthedocs.io/en/latest/>`_ to calculate
+some rudimentary properties of `stk` molecules. `stko` can be installed with
+``pip install stko``.
+
+.. GENERATED FROM PYTHON SOURCE LINES 16-28
+
+.. code-block:: Python
+
+
+    import itertools as it
+    import tempfile
+
+    import ase.io
+    import stk
+    import stko
+    from rdkit.Chem import AllChem as rdkit
+
+    import chemiscope
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 29-34
+
+Interfacing chemiscope with stk molecules
+++++++++++++++++++++++++++++++++++++++++++
+
+Generate a list of stk BuildingBlocks (representation of a molecule) with
+properties. We start by constructing a cage and host-guest complex with stk.
+
+.. GENERATED FROM PYTHON SOURCE LINES 35-60
+
+.. code-block:: Python
+
+
+    cage = stk.ConstructedMolecule(
+        topology_graph=stk.cage.FourPlusSix(
+            building_blocks=(
+                stk.BuildingBlock(
+                    smiles="NCCN",
+                    functional_groups=[stk.PrimaryAminoFactory()],
+                ),
+                stk.BuildingBlock(
+                    smiles="O=CC(C=O)C=O",
+                    functional_groups=[stk.AldehydeFactory()],
+                ),
+            ),
+            optimizer=stk.MCHammer(),
+        ),
+    )
+    host_guest = stk.ConstructedMolecule(
+        topology_graph=stk.host_guest.Complex(
+            host=stk.BuildingBlock.init_from_molecule(cage),
+            guests=stk.host_guest.Guest(
+                building_block=stk.BuildingBlock("[Br][Br]"),
+            ),
+        ),
+    )
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 61-62
+
+Including using stk to interface with rdkit molecules.
+
+.. GENERATED FROM PYTHON SOURCE LINES 63-71
+
+.. code-block:: Python
+
+
+    rdkitmol = rdkit.MolFromSmiles("Cc1ccccc1")
+    rdkitmol = rdkit.AddHs(rdkitmol)
+    rdkit.Kekulize(rdkitmol)
+    params = rdkit.ETKDGv3()
+    params.randomSeed = 0xF00D
+    rdkit.EmbedMolecule(rdkitmol, params)
+
+
+
+
+
[…]
