+++ docs/_sources/examples/sg_execution_times.rst.txt
@@ -1,0 +1,64 @@
+
+:orphan:
+
+.. _sphx_glr_examples_sg_execution_times:
+
+
+Computation times
+=================
+**00:25.865** total execution time for 10 files **from examples**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_examples_10-visualizing-biomolecules.py` (``10-visualizing-biomolecules.py``)
+     - 00:10.898
+     - 0.0
+   * - :ref:`sphx_glr_examples_11-headless.py` (``11-headless.py``)
+     - 00:09.909
+     - 0.0
+   * - :ref:`sphx_glr_examples_6-explore.py` (``6-explore.py``)
+     - 00:04.195
+     - 0.0
+   * - :ref:`sphx_glr_examples_9-stk-custom-bonds.py` (``9-stk-custom-bonds.py``)
+     - 00:00.445
+     - 0.0
+   * - :ref:`sphx_glr_examples_8-explore-with-metatomic.py` (``8-explore-with-metatomic.py``)
+     - 00:00.142
+     - 0.0
+   * - :ref:`sphx_glr_examples_5-shapes.py` (``5-shapes.py``)
+     - 00:00.133
+     - 0.0
+   * - :ref:`sphx_glr_examples_1-base.py` (``1-base.py``)
+     - 00:00.087
+     - 0.0
+   * - :ref:`sphx_glr_examples_3-trajectory.py` (``3-trajectory.py``)
+     - 00:00.024
+     - 0.0
+   * - :ref:`sphx_glr_examples_2-structure_map.py` (``2-structure_map.py``)
+     - 00:00.023
+     - 0.0
+   * - :ref:`sphx_glr_examples_4-colors.py` (``4-colors.py``)
+     - 00:00.009
+     - 0.0

[…]
