### 2026-05-02 #10380 @abhishektiwari
@shehan807 Hello, how are you tracking with your review? Let me know if anything holding up the review. You can create a new issue in the [target repository](https://github.com/lab-cosmo/chemiscope) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread here.
