+++ paper/paper.bib
@@ -7,7 +7,7 @@
   number = {51},
   pages = {2117},
   author = {Fraux, Guillaume and Cersonsky, Rose K. and Ceriotti, Michele},
-  title = {Chemiscope: interactive structure-property explorer for materials and
+  title = {{Chemiscope}: interactive structure-property explorer for materials and
   molecules},
   journal = {Journal of Open Source Software},
 }
@@ -22,7 +22,7 @@
   Rostgaard and Jakob Schiøtz and Ole Schütt and Mikkel Strange and Kristian S Thygesen
   and Tejs Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W
   Jacobsen},
-  title = {The atomic simulation environment—a Python library for working with atoms},
+  title = {The atomic simulation environment—a {Python} library for working with atoms},
   journal = {Journal of Physics: Condensed Matter},
   volume = {29},
   number = {27},
@@ -36,7 +36,7 @@
   author = {Gowers, Richard J. and Linke, Max and Barnoud, Jonathan and Reddy, Tyler J.
   E. and Melo, Manuel N. and Seyler, Sean L. and Domański, Jan and Dotson, David L. and
   Buchoux, Sébastien and Kenney, Ian M. and Beckstein, Oliver},
-  title = {MDAnalysis: A Python Package for the Rapid Analysis of Molecular Dynamics
+  title = {{MDAnalysis}: A {Python} Package for the Rapid Analysis of Molecular Dynamics
   Simulations},
   journal = {SciPy 2016},
   year = {2016},
@@ -45,7 +45,7 @@
 }
 
 @article{STK,
-  title = {stk: An extendable Python framework for automated molecular and
+  title = {stk: An extendable {Python} framework for automated molecular and
   supramolecular structure assembly and discovery},
   author = {Turcani, Lukas and Tarzia, Andrew and Szczypi{\'n}ski, Filip T. and Jelfs,
   Kim E.},
@@ -61,7 +61,7 @@
   author = {Guillaume Fraux and Len Kimms and Jonathan Fine and German P. Barletta and
   Mykola Dimura and FX Coudert and pelsa and Maximilien Levesque and Shoubhik Maiti and
   Simon Guionniere and jmintser},
-  title = {chemfiles/chemfiles: Version 0.10.4},
+  title = {{chemfiles}/{chemfiles}: Version 0.10.4},
   month = {may},
   year = {2023},
   publisher = {Zenodo},
@@ -71,7 +71,7 @@
 }
 
 @article{Mazitov2025,
-  title = {PET-MAD as a lightweight universal interatomic potential for advanced
+  title = {{PET-MAD} as a lightweight universal interatomic potential for advanced
   materials modeling},
   author = {Mazitov, Arslan and Bigi, Filippo and Kellner, Matthias and Pegolo, Paolo
   and Tisi, Davide and Fraux, Guillaume and Pozdnyakov, Sergey and Loche, Philip and
@@ -100,7 +100,7 @@
 }
 
 @article{metatensor,
-  title = {Metatensor and metatomic: Foundational libraries for interoperable atomistic
+  title = {{metatensor} and {metatomic}: Foundational libraries for interoperable atomistic
   machine learning},
   author = {Filippo Bigi and Joseph W. Abbott and Philip Loche and Arslan Mazitov and
   Davide Tisi and Marcel F. Langer and Alexander Goscinski and Paolo Pegolo and Sanggyu
@@ -119,7 +119,7 @@
 @article{Jupyter,
   author = {Granger, Brian E. and Pérez, Fernando},
   journal = {Computing in Science & Engineering},
-  title = {Jupyter: Thinking and Storytelling With Code and Data},
[…]
