+++ paper/paper.bib
@@ -472,21 +472,21 @@
 @misc{MaterialsCloudChemiscopeSearch,
   title = {Materials Cloud Archive},
   url = {https://archive.materialscloud.org/search?q=&f=ext_apps%3Achemiscope&l=list&p=1&s=10&sort=newest},
-  note = {Accessed 2026-02-16},
+  note = {Accessed 2026-05-14},
   year = {2026},
 }
 
 @misc{PyPIStatsChemiscope,
   title = {{PyPI} Stats for chemiscope},
   url = {https://pypistats.org/packages/chemiscope},
-  note = {Accessed 2026-02-16},
+  note = {Accessed 2026-05-14},
   year = {2026},
 }
 
 @misc{AtomisticCookbook,
   title = {The Atomistic Cookbook},
   url = {https://atomistic-cookbook.org/software/chemiscope.html},
-  note = {Accessed 2026-02-16},
+  note = {Accessed 2026-05-14},
   year = {2026},
 }
 
