+++ paper/paper.bib
@@ -100,8 +100,8 @@
 }
 
 @article{metatensor,
-  title = {{metatensor} and {metatomic}: Foundational libraries for interoperable atomistic
-  machine learning},
+  title = {{metatensor} and {metatomic}: Foundational libraries for interoperable
+  atomistic machine learning},
   author = {Filippo Bigi and Joseph W. Abbott and Philip Loche and Arslan Mazitov and
   Davide Tisi and Marcel F. Langer and Alexander Goscinski and Paolo Pegolo and Sanggyu
   Chong and Rohit Goswami and Pol Febrer and Sofiia Chorna and Matthias Kellner and
