### 2026-05-14 3009da4
update Jorgensen et al. ref as published in a journal
