+++ paper/paper.bib
@@ -440,12 +440,13 @@
   title = {Interpretable Visualizations of Data Spaces for Classification Problems},
   author = {Jorgensen, Christian and Lin, Arthur Y. and Vasavada, Rhushil and Cersonsky,
   Rose K.},
-  year = {2025},
-  eprint = {2503.05861},
-  archivePrefix = {arXiv},
-  primaryClass = {cs.LG},
-  doi = {10.48550/arXiv.2503.05861},
-  url = {https://arxiv.org/abs/2503.05861},
+  journal = {Machine Learning: Science and Technology},
+  volume = {7},
+  number = {2},
+  pages = {025008},
+  year = {2026},
+  doi = {10.1088/2632-2153/ae466e},
+  url = {https://doi.org/10.1088/2632-2153/ae466e},
 }
 
 @article{MD22,
