### 2026-05-18 c280396
Add LOD to docs (#557)
added mentioning of the LOD in the docs following
https://github.com/lab-cosmo/chemiscope/issues/556.
