### 2026-05-19 f03a2cf
Use best-min-hash strategy to pick structures in the LOD routine (#558)
Found it, fixed it.
