+++ python/examples/12-parameterized-properties.py
@@ -1,0 +1,72 @@
+"""
+Parameterized properties
+========================
+
+This example demonstrates the use of parameterized (multidimensional) properties
+in chemiscope. These are displayed as curves in the info panel.
+
+We compute the radial distribution function (RDF) for each structure and display
+it as a parameterized property alongside the structure viewer.
+"""
+
+# %%
+#
+
+import ase.io
+import numpy as np
+from ase.geometry.analysis import Analysis
+
+import chemiscope
+
+
+# %%
+#
+# Load structures
+
+structures = ase.io.read("data/explore_c-gap-20u.xyz", ":")
+
+# %%
+#
+# Compute the RDF for each structure using ASE
+
+rmax = 3.5
+nbins = 50
+rdf_values = []
+for structure in structures:
+    ana = Analysis(structure)
+    rdf = ana.get_rdf(rmax=rmax, nbins=nbins, return_dists=False)[0]
+    rdf_values.append(rdf)
+
+# The r grid corresponding to the RDF bins
+dr = rmax / nbins
+r_grid = np.linspace(dr / 2, rmax - dr / 2, nbins)
+
+# %%
+#
+# Create a chemiscope widget with the RDF as a parameterized property
+
+chemiscope.show(
+    structures=structures,
+    properties={
+        "RDF": {
+            "target": "structure",
+            "values": np.array(rdf_values),
+            "parameters": ["r"],
+            "description": "Radial distribution function g(r)",
+        },
+        "config_type": [s.info["config_type"] for s in structures],
+    },
+    parameters={
+        "r": {
+            "values": r_grid,
+            "units": "Angstrom",
+        },
+    },
+    mode="structure",
+    metadata=dict(
+        name="RDF of selected C structures",
+        description="Radial distribution functions computed for carbon structures "
+        "from the C-GAP-20U dataset, displayed as parameterized properties.",
+    ),
+    settings=chemiscope.quick_settings(periodic=True),
+)
