### 2026-05-21 ab59d4e
Make chemiscope shapes obey transparency
