+++ src/structure/shapes.ts
@@ -338,6 +338,46 @@ function isPositiveInteger(o: unknown): boolean {
     return typeof o === 'number' && Number.isInteger(o) && o >= 0;
 }
 
+// Matches a CSS-style `#RRGGBBAA` hex string. Anchored so 6-digit `#RRGGBB`,
+// named colors, and rgb()/rgba() strings fall through unchanged to 3Dmol.
+const HEX_RGBA_STRING = /^#([0-9a-f]{6})([0-9a-f]{2})$/i;
+
+// Extract the alpha component from a ColorSpec. 3Dmol's Color parser is
+// RGB-only, so any alpha would otherwise be dropped silently. Alpha is
+// recognized in two forms:
+//   - object color `{r, g, b, a}` with `a` in [0, 1];
+//   - 8-digit CSS hex string `#RRGGBBAA` (alpha in the last two hex digits).
+// All other colors (named, `#RRGGBB`, numeric, rgb(), etc.) are opaque.
+function colorAlpha(color: ColorSpec | undefined): number {
+    if (color === undefined || color === null) {
+        return 1;
+    }
+    if (typeof color === 'object' && typeof color.a === 'number') {
+        return Math.max(0, Math.min(1, color.a));
+    }
+    if (typeof color === 'string') {
+        const match = HEX_RGBA_STRING.exec(color);
+        if (match) {
+            return parseInt(match[2], 16) / 255;
+        }
+    }
+    return 1;
+}
+
+// Return a ColorSpec stripped of any alpha component so 3Dmol's RGB-only
+// parser sees the right RGB. `#RRGGBBAA` strings are truncated to `#RRGGBB`;
+// object colors with a spurious `a` field are passed through (3Dmol ignores
+// extra keys); everything else is unchanged.
+function stripAlpha(color: ColorSpec | undefined): ColorSpec | undefined {
+    if (typeof color === 'string') {
+        const match = HEX_RGBA_STRING.exec(color);
+        if (match) {
+            return '#' + match[1];
+        }
+    }
+    return color;
+}
+
 export class Sphere extends Shape {
     public radius: number;
 
@@ -1137,6 +1177,36 @@ export class CustomShape extends Shape {
 
 const MAX_VERTICES_IN_SINGLE_SHAPE = 32768;
 
+/**
+ * Submit the accumulated merged shape to the viewer and reset the in-place
+ * buffers so further shapes can keep accumulating with a fresh opacity. Safe
+ * to call when the merged shape is empty.
+ */
+export function flushMergedShapes(
+    mergedShapes: $3Dmol.CustomShapeSpec,
+    viewer: $3Dmol.GLViewer
+): void {
+    if (!Array.isArray(mergedShapes.faceArr) || mergedShapes.faceArr.length === 0) {
+        return;
+    }
+    viewer.addCustom(mergedShapes);
+    if (Array.isArray(mergedShapes.vertexArr)) {
+        mergedShapes.vertexArr.length = 0;
+    }
+    if (Array.isArray(mergedShapes.normalArr)) {
+        mergedShapes.normalArr.length = 0;
+    }
+    if (Array.isArray(mergedShapes.faceArr)) {
+        mergedShapes.faceArr.length = 0;
+    }
+    if (Array.isArray(mergedShapes.color)) {
+        mergedShapes.color.length = 0;
+    }
+    // 3Dmol only supports one opacity per shape, so each flushed batch
+    // gets its own — clear it so the next batch can set a fresh value.
+    mergedShapes.opacity = undefined;
+}
+
 export function mergeShapes(
     mergedShapes: $3Dmol.CustomShapeSpec,
     shape: $3Dmol.CustomShapeSpec,
[…]
