### 2026-05-21 4d9a73d
Update python/examples/12-parameterized-properties.py
Co-authored-by: Sofiia Chorna <71195115+sofiia-chorna@users.noreply.github.com>
