### 2026-05-21 0aa6097
Add transparency example
