### 2026-05-21 f8dd2e4
Make chemiscope shapes obey alpha specifications (#559)
