### 2026-05-26 351f145
update features to feature everywhere, use sample_kind instead of per_atom
