### 2026-05-28 00d63fe
implement streaming dataset loading
