+++ app/streaming.ts
@@ -1,0 +1,726 @@
+import type { Dataset, Environment, Property, Structure, UserStructure } from '../src/dataset';
+
+// structures per IndexedDB record
+const BATCH_SIZE = 256;
+
+// batch writes kept in flight
+const MAX_INFLIGHT_WRITES = 32;
+
+// only drop the consumed buffer prefix past this size
+const COMPACT_THRESHOLD = 1 << 16;
+
+// unique id for the store name
+function uid(): string {
+    if (typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function') {
+        return crypto.randomUUID();
+    }
+    return Math.random().toString(36).slice(2) + Date.now().toString(36);
+}
+
+export interface StreamingDataset {
+    // dataset with lightweight structure placeholders (`{ size, data: index }`)
+    dataset: Dataset;
+    // on-demand structure loader
+    loadStructure: (index: number) => Promise<Structure>;
+    // release the backing IndexedDB store
+    release: () => Promise<void>;
+}
+
+export interface StreamingProgress {
+    // compressed bytes read so far
+    bytesRead: number;
+    // total compressed size, if known
+    bytesTotal: number;
+    // structures stored so far
+    structures: number;
+}
+
+// parse JSON, tolerating Python's bare `NaN`
+export function parseJsonWithNaN(text: string): unknown {
+    return JSON.parse(text.replace(/\bNaN\b/g, '"***NaN***"'), (_key, value: unknown) =>
+        value === '***NaN***' ? NaN : value
+    ) as unknown;
+}
+
+// like Number(token) but throws on non-numbers instead of silently returning
+// 0 (for "") or NaN
+function parseNumberToken(token: string): number {
+    if (token === 'NaN') {
+        return NaN;
+    }
+    const value = Number(token);
+    if (token === '' || Number.isNaN(value)) {
+        throw new Error(`invalid number in dataset: ${JSON.stringify(token)}`);
+    }
+    return value;
+}
+
+// raw structure JSON in IndexedDB, batched, with a one-batch read cache
+class StructureStore {
+    // db names are `${NAME_PREFIX}${timestamp}-${uid}`
+    private static readonly NAME_PREFIX = 'chemiscope-stream-';
+    private static readonly ORPHAN_MAX_AGE_MS = 60 * 60 * 1000; // 1 hour
+
+    private db: IDBDatabase;
+    private readonly storeName = 'structures';
+
+    // which batch is currently held in `cache` (-1 = none)
+    private cacheBatch = -1;
+    private cache: string[] = [];
+
+    private constructor(db: IDBDatabase) {
+        this.db = db;
+    }
+
+    public static async open(): Promise<StructureStore> {
+        if (typeof indexedDB === 'undefined') {
+            throw new Error('IndexedDB is not available in this environment');
+        }
+
+        // reap leftovers from crashed sessions before adding ours
+        await StructureStore.sweepOrphans();
+
+        // unique name so back-to-back loads never collide
+        const dbName = `${StructureStore.NAME_PREFIX}${Date.now()}-${uid()}`;
+
+        const db = await new Promise<IDBDatabase>((resolve, reject) => {
+            const req = indexedDB.open(dbName, 1);
+            req.onupgradeneeded = () => req.result.createObjectStore('structures');
+            req.onsuccess = () => resolve(req.result);
+            req.onerror = () => reject(req.error ?? new Error('failed to open IndexedDB'));
+        });
+
+        return new StructureStore(db);
+    }
+
+    // drop stores left behind by sessions that never called release()
+    private static async sweepOrphans(): Promise<void> {
+        const factory = indexedDB as IDBFactory & {
+            databases?: () => Promise<{ name?: string }[]>;
+        };
[…]
