### 2026-06-02 eb0236e
use journal ref for MD22 citation
