+++ paper/paper.bib
@@ -452,15 +452,16 @@
 @article{MD22,
   title = {Accurate global machine learning force fields for molecules with hundreds of
   atoms},
-  author = {Chmiela, Stefan and Vassilev-Galindo, Valentin and Unke, Oliver T. and
-  Kabylda, Adil and Sauceda, Huziel E. and Tkatchenko, Alexandre and Müller,
-  Klaus-Robert},
-  year = {2022},
-  eprint = {2209.14865},
-  archivePrefix = {arXiv},
-  primaryClass = {physics.chem-ph},
-  doi = {10.48550/arXiv.2209.14865},
-  url = {https://arxiv.org/abs/2209.14865},
+  author = {Stefan Chmiela and Valentin Vassilev-Galindo and Oliver T. Unke and Adil
+  Kabylda and Huziel E. Sauceda and Alexandre Tkatchenko and Klaus-Robert Müller},
+  year = {2023},
+  journal = {Science Advances},
+  volume = {9},
+  number = {2},
+  pages = {eadf0873},
+  doi = {10.1126/sciadv.adf0873},
+  url = {https://www.science.org/doi/abs/10.1126/sciadv.adf0873},
+  eprint = {https://www.science.org/doi/pdf/10.1126/sciadv.adf0873},
 }
 
 @misc{plotlyjs,
