### 2026-06-02 64a4356
implement streaming dataset loading and parsing
