+++ app/streaming/index.ts
@@ -1,0 +1,413 @@
+import type { Dataset, Environment, Property, Structure, UserStructure } from '../../src/dataset';
+import { BATCH_SIZE, StructureStore } from './store';
+import { StreamingJSONParser, parseJsonWithNaN } from './parser';
+
+export { parseJsonWithNaN };
+
+const MAX_PENDING_WRITES = 32;
+const COMPACT_THRESHOLD = 1 << 16;
+const PROGRESS_THROTTLE_MS = 100;
+
+export interface StreamingDataset {
+    dataset: Dataset;
+    loadStructure: (index: number) => Promise<Structure>;
+    release: () => Promise<void>;
+}
+
+export interface StreamingProgress {
+    bytesRead: number;
+    bytesTotal: number;
+    structures: number;
+    phase:
+        | 'structures'
+        | 'flushing'
+        | 'properties'
+        | 'environments'
+        | 'finalizing'
+        | 'preparing';
+}
+
+const WHITESPACE = new Set([' ', '\t', '\n', '\r']);
+
+function extractSize(raw: string, index: number): number {
+    const missing = () => new Error(`structure ${index} is missing an integer "size"`);
+    let depth = 0;
+    let inString = false;
+    let escape = false;
+    let expectKey = false;
+
+    for (let i = 0; i < raw.length; i++) {
+        const c = raw[i];
+        if (inString) {
+            if (escape) {
+                escape = false;
+            } else if (c === '\\') {
+                escape = true;
+            } else if (c === '"') {
+                inString = false;
+            }
+            continue;
+        }
+        if (c === '"') {
+            if (depth === 1 && expectKey && raw.startsWith('"size"', i)) {
+                let j = i + 6;
+                while (j < raw.length && WHITESPACE.has(raw[j])) {
+                    j++;
+                }
+                if (raw[j] !== ':') {
+                    throw missing();
+                }
+                j++;
+                while (j < raw.length && WHITESPACE.has(raw[j])) {
+                    j++;
+                }
+                const start = j;
+                while (j < raw.length && raw[j] >= '0' && raw[j] <= '9') {
+                    j++;
+                }
+                if (j === start || raw[j] === '.' || raw[j] === 'e' || raw[j] === 'E') {
+                    throw missing();
+                }
+                return parseInt(raw.slice(start, j), 10);
+            }
+            inString = true;
+            expectKey = false;
+        } else if (c === '{') {
+            depth++;
+            expectKey = depth === 1;
+        } else if (c === '[') {
+            depth++;
+            expectKey = false;
+        } else if (c === '}' || c === ']') {
+            depth--;
+            expectKey = false;
+        } else if (c === ',' && depth === 1) {
+            expectKey = true;
+        } else if (c === ':' && depth === 1) {
+            expectKey = false;
+        }
+    }
+    throw missing();
+}
+
+/**
+ * Stream a chemiscope dataset from a File: the huge structures array is
+ * offloaded to browser storage and replaced in the returned dataset with
+ * lightweight placeholders, while everything else is parsed in memory.
+ */
+export async function loadDatasetStreaming(
+    file: File,
+    onProgress?: (progress: StreamingProgress) => void
+): Promise<StreamingDataset> {
+    const isGzipped = file.name.endsWith('.gz');
+    if (isGzipped && typeof DecompressionStream === 'undefined') {
+        throw new Error(
+            'streaming load of gzipped files requires DecompressionStream (unavailable in this browser)'
+        );
+    }
+
+    // count compressed bytes so progress is against file.size
+    let bytesRead = 0;
+    const counting = new TransformStream<Uint8Array, Uint8Array>({
+        transform(chunk, controller) {
[…]
