### 2026-06-03 c774650
Prioritize selected points when doing LOD to avoid artifacts when the selection is very small
