### 2026-06-03 5e71b0a
Fix a signed/unsigned int bug
