### 2026-06-03 b072390
Prioritize selected points when doing LOD to avoid artifacts (#562)
