### 2026-06-04 b5b0cb7
remove unused worker and parser files
