+++ app/streaming.ts
@@ -1,0 +1,632 @@
+import type { Dataset, Environment, Property, Structure, UserStructure } from '../src/dataset';
+
+// how many structures go into a single store record
+const BATCH_SIZE = 2048;
+// how many store writes can run in the background before we wait
+const MAX_PENDING_WRITES = 32;
+// drop the consumed part of the parse buffer past this size
+const COMPACT_THRESHOLD = 1 << 16;
+// progress updates are throttled to keep the displayed numbers readable
+const PROGRESS_THROTTLE_MS = 100;
+
+const WHITESPACE = new Set([' ', '\t', '\n', '\r']);
+
+export interface StreamingDataset {
+    dataset: Dataset;
+    loadStructure: (index: number) => Promise<Structure>;
+    release: () => Promise<void>;
+}
+
+export interface StreamingProgress {
+    bytesRead: number;
+    bytesTotal: number;
+    structures: number;
+    phase:
+        | 'structures'
+        | 'flushing'
+        | 'properties'
+        | 'environments'
+        | 'finalizing'
+        | 'preparing';
+}
+
+export function parseJsonWithNaN(text: string): unknown {
+    if (!text.includes('NaN')) {
+        return JSON.parse(text) as unknown;
+    }
+
+    let out = '';
+    let chunkStart = 0;
+    let inString = false;
+    let escape = false;
+    for (let i = 0; i < text.length; i++) {
+        const c = text[i];
+        if (inString) {
+            if (escape) {
+                escape = false;
+            } else if (c === '\\') {
+                escape = true;
+            } else if (c === '"') {
+                inString = false;
+            }
+            continue;
+        }
+        if (c === '"') {
+            inString = true;
+            continue;
+        }
+        if (c === 'N' && text[i + 1] === 'a' && text[i + 2] === 'N') {
+            out += text.slice(chunkStart, i) + '"***NaN***"';
+            i += 2;
+            chunkStart = i + 1;
+        }
+    }
+    out += text.slice(chunkStart);
+    return JSON.parse(out, (_key, value: unknown) =>
+        value === '***NaN***' ? NaN : value
+    ) as unknown;
+}
+
+function makeDepthScanner(): (buf: string, from: number) => number {
+    let depth = 0;
+    let inString = false;
+    let escape = false;
+    let started = false;
+
+    return function scan(buf: string, from: number): number {
+        for (let i = from; i < buf.length; i++) {
+            const char = buf[i];
+
+            if (inString) {
+                if (escape) {
+                    escape = false;
+                } else if (char === '\\') {
+                    escape = true;
+                } else if (char === '"') {
+                    inString = false;
+                }
+                continue;
+            }
+
+            if (char === '"') {
+                inString = true;
+            } else if (char === '{' || char === '[') {
+                depth++;
+                started = true;
+            } else if (char === '}' || char === ']') {
+                depth--;
+                if (started && depth === 0) {
+                    return i + 1;
+                }
+            }
+        }
+        return -1;
+    };
+}
+
+function extractSize(raw: string, index: number): number {
+    let inString = false;
+    let escape = false;
+
+    for (let i = 0; i < raw.length; i++) {
+        const c = raw[i];
+        if (inString) {
+            if (escape) {
+                escape = false;
+            } else if (c === '\\') {
+                escape = true;
+            } else if (c === '"') {
+                inString = false;
+            }
+            continue;
+        }
+        if (c !== '"') {
+            continue;
+        }
+        if (!raw.startsWith('"size"', i)) {
+            inString = true;
+            continue;
+        }
+
+        let j = i + 6;
+        while (j < raw.length && WHITESPACE.has(raw[j])) {
+            j++;
+        }
+        if (raw[j] !== ':') {
[…]
