### 2026-06-06 ad01e50
Move as much code as possible into streaming.ts
