### 2026-06-06 75836df
Added a docstring for streaming.ts
