### 2026-06-06 e4a1fb9
implement streaming dataset loading in the app (#561)
