### 2026-06-06 95a73ce
Add CITATION.cff for JOSS 1.0 paper; cite it as primary; bump to v1.0.4
