### 2026-06-06 f9fb275
Fix a bunch of problems related to the partial updates of the visualization state following changes in selection
