### 2026-06-08 89bb528
Fix selection range update in jupyter
Also found that the axes range updates had been broken since forever, so fixed those as well
