### 2026-06-08 f398e84
Fix a bunch of problems related to the partial updates of the selction (#563)
Color scale and LOD changes were not reflected in the visualization
following changes of the subselection
