+++ src/map/map.ts
@@ -12,7 +12,6 @@ import * as plotlyStyles from './plotly/plotly-styles';
 import { Property, Settings } from '../dataset';
 
 import { DisplayTarget, EnvironmentIndexer, Indexes } from '../indexer';
-import { OptionModificationOrigin } from '../options';
 import { cameraToPlotly, plotlyToCamera } from '../utils/camera';
 import {
     Bounds,
@@ -97,10 +96,11 @@ export class PropertiesMap {
     private static readonly LOD_THRESHOLD = 50000;
     /// Stores the subset of point indices to display when LOD is active
     private _lodIndices: number[] | null = null;
-    /**
-     * Guard to skip concurrent LOD updates. When set, other callers simply return early
-     * instead of waiting
-     */
+    /// True while writing derived option values programmatically; the reactive
+    /// onchange handlers (updateColors, axis rangeChange) skip while it is set.
+    private _internalUpdate = false;
+    /// Guard to skip concurrent LOD updates. When set, other callers simply
+    // return early instead of waiting
     private _lodBusy = false;
     // Timeout id used to batch plotly afterplot events
     private _afterplotRequest: number | null = null;
@@ -727,10 +727,9 @@ export class PropertiesMap {
         // function creating a function to be used as onchange callback
         // for <axis>.min and <axis>.max
         const rangeChange = (name: string, axis: AxisOptions, minOrMax: 'min' | 'max') => {
-            return (_: number, origin: OptionModificationOrigin) => {
-                if (origin === 'JS') {
-                    // prevent recursion: this function calls relayout, which then
-                    // calls _afterplot, which reset the min/max values.
+            return () => {
+                // ignore programmatic (derived) min/max writes
+                if (this._internalUpdate) {
                     return;
                 }
                 const min = axis.min.value;
@@ -739,11 +738,14 @@ export class PropertiesMap {
                     this.warnings.sendMessage(
                         `The inserted min and max values in ${name} are such that min > max! The last inserted value was reset.`
                     );
-                    if (minOrMax === 'min') {
-                        axis.min.reset();
-                    } else {
-                        axis.max.reset();
-                    }
+                    // revert the invalid value without re-running this handler
+                    this.withInternalUpdate(() => {
+                        if (minOrMax === 'min') {
+                            axis.min.reset();
+                        } else {
+                            axis.max.reset();
+                        }
+                    });
                     return;
                 }
 
@@ -794,8 +796,10 @@ export class PropertiesMap {
 
         this._options.z.property.onchange.push(() => {
             // Reset min/max when changing property to trigger autoscale
-            this._options.z.min.value = NaN;
-            this._options.z.max.value = NaN;
+            this.withInternalUpdate(() => {
+                this._options.z.min.value = NaN;
+                this._options.z.max.value = NaN;
+            });
             negativeLogWarning(this._options.z);
 
             const was3D = this._plot._fullData[0].type === 'scatter3d';
@@ -824,8 +828,10 @@ export class PropertiesMap {
                 .then(() => {
                     // The zrange is now known, and we can trigger a proper subsampling
                     const zRange = this._plot._fullLayout.scene.zaxis.range as number[];
-                    this._options.z.min.value = zRange[0];
-                    this._options.z.max.value = zRange[1];
+                    this.withInternalUpdate(() => {
+                        this._options.z.min.value = zRange[0];
+                        this._options.z.max.value = zRange[1];
+                    });
 
                     if (this._is3D()) {
                         this._setScaleStep(this._getBounds().z as number[], 'z');
@@ -888,6 +894,15 @@ export class PropertiesMap {
[…]
