+++ tox.ini
@@ -15,7 +15,6 @@ lint_folders = "{toxinidir}/python" "{toxinidir}/setup.py"
 [testenv:build]
 description = Build the wheel for chemiscope (internal)
 deps =
-    jupyterlab
     setuptools
     wheel
 
