### 2026-06-18 e05efbb
Fixed a bunch of layout errors
