+++ src/map/map.ts
@@ -111,6 +111,12 @@ export class PropertiesMap {
     private _mouseupHandler?: () => void;
     // last event data received from plotly_relayouting
     private _lastRelayoutingEvent: Plotly.PlotRelayoutEvent | null = null;
+    // pending axis-range relayout, coalesced to one relayout per tick
+    private _pendingRangeUpdate: Record<string, unknown> = {};
+    private _rangeFlushScheduled = false;
+    // observes the container so the plot reflows on non-window resizes
+    private _resizeObserver?: ResizeObserver;
+    private _resizePending = false;
 
     /**
      * Create a new {@link PropertiesMap} inside the DOM element with the given HTML
@@ -217,12 +223,26 @@ export class PropertiesMap {
             plotlyStyles.globalStyleSheet,
             plotlyStyles.getPlotStyleSheet(this._plot),
         ];
+
+        // Plotly's `responsive` config only reacts to window resizes; observe the
+        // container so the plot also reflows when its size changes for other reasons
+        // (e.g. toggling a Jupyter sidebar).
+        this._resizeObserver = new ResizeObserver(() => {
+            if (this._resizePending) {
+                return;
+            }
+            this._resizePending = true;
+            window.requestAnimationFrame(() => {
+                this._resizePending = false;
+                Plotly.Plots.resize(this._plot);
+            });
+        });
+        this._resizeObserver.observe(this._plot);
     }
 
-    /**
-     * Remove all HTML added by this {@link PropertiesMap} in the current document
-     */
     public remove(): void {
+        this._resizeObserver?.disconnect();
+
         // Remove the the shadow root's host. It is not possible to remove the shadow root directly.
         this._shadow.host.remove();
 
@@ -449,7 +469,6 @@ export class PropertiesMap {
         // Get config
         const config = this._getConfig();
 
-        // Create an empty plot and fill it below
         Plotly.newPlot(this._plot, traces, layout, config)
             .then(() => {
                 // In some cases (e.g. in Jupyter notebooks) plotly does not comply
@@ -669,7 +688,7 @@ export class PropertiesMap {
             this._computeLOD(bounds);
             // Force a full restyle. Since _lodIndices will be null if disabled,
             // this will render all points.
-            this._restyleFull();
+            void this._restyleFull();
         });
 
         // ======= camera state update
@@ -732,6 +751,10 @@ export class PropertiesMap {
                 if (this._internalUpdate) {
                     return;
                 }
+                // the z axis only exists in 3D; relaying it out in 2D makes Plotly throw
+                if (name === 'zaxis' && !this._is3D()) {
+                    return;
+                }
                 const min = axis.min.value;
                 const max = axis.max.value;
                 if (min > max) {
@@ -751,14 +774,18 @@ export class PropertiesMap {
 
                 negativeLogWarning(axis);
 
-                if (this._is3D()) {
-                    this._relayout({
-                        [`scene.${name}.range`]: [min, max],
-                        [`scene.${name}.autorange`]: false,
-                    } as unknown as Layout);
+                const key = this._is3D() ? `scene.${name}` : name;
+                let update: Record<string, unknown>;
+                if (isNaN(min) && isNaN(max)) {
+                    // both bounds cleared: turn autorange on. A `[NaN, NaN]` range is
+                    // read by Plotly as autorange-off with default bounds
+                    update = { [`${key}.autorange`]: true };
+                } else if (this._is3D()) {
+                    update = { [`${key}.range`]: [min, max], [`${key}.autorange`]: false };
                 } else {
-                    this._relayout({ [`${name}.range`]: [min, max] });
+                    update = { [`${key}.range`]: [min, max] };
                 }
[…]
