+++ python/chemiscope/widget.py
@@ -1,0 +1,828 @@
+# -*- coding: utf-8 -*-
+import asyncio
+import base64
+import gzip
+import itertools
+import json
+import warnings
+from pathlib import Path
+
+import anywidget
+from traitlets import Bool, Dict, Int, Unicode
+
+from .input import create_input
+
+
+# Single ES module loaded by anywidget for all three widget modes. It is built by
+# `npm run build:anywidget` from python/widget/src/anywidget.ts.
+_ESM_PATH = Path(__file__).parent / "static" / "chemiscope-widget.mjs"
+
+
+class ChemiscopeWidgetBase(anywidget.AnyWidget):
+    _esm = _ESM_PATH
+
+    # widget mode, read by the JS side to pick which view to render; overridden by
+    # each subclass
+    mode = Unicode("default").tag(sync=True)
+
+    value = Unicode().tag(sync=True)
+    has_metadata = Bool().tag(sync=True)
+
+    # synchronized settings from the JS side. You can assign to this field to
+    # change what's being displayed by chemiscope, but you need to assign a full
+    # dictionary (`widget.settings["map"]["x"]["property"] = "foo"` will not
+    # work, but `widget.settings = updated_settings` will).
+    settings = Dict().tag(sync=True)
+    # timeout for warning messages
+    warning_timeout = Int(10000).tag(sync=True)
+    # currently selected environment/structure/atom
+    selected_ids = Dict().tag(sync=True)
+    # index of the currently active viewer in the grid
+    active_viewer = Int(0).tag(sync=True)
+
+    def __init__(
+        self, data, has_metadata, warning_timeout=10000, cache_structures=True
+    ):
+        super().__init__()
+        self.has_metadata = has_metadata
+        if "settings" in data:
+            self.settings = data["settings"]
+
+        # timeout for warning messages (ms). 0 to make persistent, -1 to disable
+        self.warning_timeout = warning_timeout
+
+        self._request_counter = itertools.count()
+        self._pending_requests = {}
+
+        # hold structures on the python side to save js memory
+        self._structures_cache = {}
+        if cache_structures:
+            data["structures"] = self._cache_structures(data["structures"])
+
+        # pass data to javascript through a traitlet
+        self.value = json.dumps(data)
+
+        # listen for custom messages from the JS side
+        self.on_msg(self._handle_js_msg)
+
+    def _cache_structures(self, structures):
+        """Cache full structures and return structure references"""
+        if not structures or "data" in structures[0]:
+            return structures
+
+        structure_refs = []
+        for i, structure in enumerate(structures):
+            if "data" in structure:
+                raise ValueError("Got a mix of explicit and external structures")
+
+            cache_key = f"structure-{i}"
+            self._structures_cache[cache_key] = structure
+            structure_refs.append({"size": structure["size"], "data": cache_key})
+
+        return structure_refs
+
+    def get_map_image(self):
+        """
+        Request a snapshot of the map. Returns a Future that resolves to the
+        image data (PNG formatted).
+
+        This method is asynchronous. In a notebook, you should ``await``
+        it to get the data.
+
+        .. code-block:: python
+
+            # Get raw image data
+            data = await widget.get_map_image()
+
+            # Display it
+            from IPython.display import Image
+
+            display(Image(data))
+
+        :return: A Future that resolves to the image data as bytes.
+        """
+
+        if self.mode == "structure":
+            raise RuntimeError(
+                "Cannot retrieve map image: this widget is a structure-only viewer."
+            )
+
+        return self._request_screenshot("map")
+
+    def get_structure_image(self):
+        """
+        Request a snapshot of the active structure viewer.
+        Returns a Future that resolves to the image data (PNG formatted).
+
[…]
