### 2026-06-19 fe5a983
Jupyter shims again
