### 2026-06-20 135bbd6
More readable if-then-else
