### 2026-06-20 e8b363d
Style properly the warning pop-up
