### 2026-06-24 5ec4778
Fix vscode problem, and correct a truncated docstring
