### 2026-06-24 57e101f
Anywidget rewrite of the chemiscope widget (#579)
TBD whether this should be cleaned and merged, or just used for a more
manual rewrite inspired by it

---------

Co-authored-by: Sofiia Chorna <71195115+sofiia-chorna@users.noreply.github.com>
