+++ python/widget/src/widget.ts
@@ -1,11 +1,14 @@
-import { DOMWidgetView } from '@jupyter-widgets/base';
-import { JSONValue } from '@lumino/coreutils';
+import type { AnyModel } from '@anywidget/types';
 import Plausible from 'plausible-tracker';
 
 import { Warnings, binarySearch, generateGUID, getByID } from '../../../src/utils';
 
-// Import the CSS
-import './widget.css';
+// Widget layout stylesheet, imported as a constructable stylesheet so it can be
+// adopted into the root node that actually contains the widget. Some hosts (e.g.
+// marimo) render the widget inside a shadow root, where a <style> injected into
+// document.head (the default css-loader behaviour) would not apply, breaking the
+// grid layout and collapsing the structure panel.
+import widgetStyles from '!css-loader?{"exportType":"css-style-sheet","url":false}!./widget.css';
 
 import {
     DefaultConfig,
@@ -45,7 +48,16 @@ interface StructureSequenceRequest {
     settings?: Partial<Settings>[];
 }
 
-class ChemiscopeBaseView extends DOMWidgetView {
+// Loosely-typed handler for model events, so that handlers with narrow argument
+// types can still be tracked for removal on dispose.
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type ModelEventHandler = (...args: any[]) => void;
+
+class ChemiscopeBaseView {
+    // anywidget model and DOM container, provided by the `render` entry point
+    protected model: AnyModel;
+    protected el: HTMLElement;
+
     protected visualizer?: DefaultVisualizer | StructureVisualizer | MapVisualizer;
     protected guid!: string;
     protected warnings: Warnings = new Warnings();
@@ -60,26 +72,96 @@ class ChemiscopeBaseView extends DOMWidgetView {
         { resolve: (s: Structure) => void; reject: (e: Error) => void }
     >();
 
+    // model event handlers, tracked so they can be removed in `dispose`
+    private _modelListeners: { event: string; handler: ModelEventHandler }[] = [];
+
+    // roots into which the widget stylesheet has already been adopted, to avoid
+    // adding it more than once per document/shadow root
+    private static _styledRoots = new WeakSet<Document | ShadowRoot>();
+
     // Flag to prevent infinite loops when updating settings from Python
     protected _updatingFromPython = false;
 
+    constructor(model: AnyModel, el: HTMLElement) {
+        this.model = model;
+        this.el = el;
+    }
+
+    /** Register a model event handler and track it for removal on dispose */
+    protected _listen(event: string, handler: ModelEventHandler): void {
+        this.model.on(event, handler);
+        this._modelListeners.push({ event, handler });
+    }
+
     public render(): void {
+        this._adoptStyles();
+
         PlausibleTracker.trackPageview({
             url: (location.pathname.split('/')[1] || '') + '/' + this.getClassName(),
         });
 
         this.guid = `chsp-${generateGUID()}`;
 
-        this.model.on('change:warning_timeout', () => this._updateWarningTimeout());
+        this._listen('change:warning_timeout', () => this._updateWarningTimeout());
         this._enableScreenshotMessages();
     }
 
-    public remove(): unknown {
+    /**
+     * Adopt the widget layout stylesheet into the root node containing this
+     * widget. Using `adoptedStyleSheets` (rather than a document.head <style>)
+     * ensures the layout CSS applies even when the host renders the widget inside
+     * a shadow root, as marimo does.
+     */
+    private _adoptStyles(): void {
+        // Only Document and ShadowRoot support `adoptedStyleSheets`. Some hosts
+        // (e.g. JupyterLab) call render before the element is attached, so
[…]
