+++ setup.py
@@ -1,11 +1,9 @@
 #!/usr/bin/env python3
 import glob
-import json
 import os
 import shutil
 import subprocess
 import sys
-import warnings
 
 from setuptools import setup
 from setuptools.command.bdist_egg import bdist_egg
@@ -30,7 +28,7 @@ class bdist_egg_disabled(bdist_egg):
 NPM_BUILD_INPUT = [
     # source files
     *glob.glob("src/**/*", recursive=True),
-    *glob.glob("python/jupyter/src/**/*", recursive=True),
+    *glob.glob("python/widget/src/**/*", recursive=True),
     # dependencies
     "package.json",
     "package-lock.json",
@@ -42,9 +40,8 @@ NPM_BUILD_INPUT = [
 ]
 
 NPM_BUILD_OUTPUT = [
-    # jupyter extensions
-    "python/jupyter/nbextension/chemiscope.min.js",
-    "python/jupyter/labextension/package.json",
+    # anywidget ES module for the notebook widget
+    "python/chemiscope/static/chemiscope-widget.mjs",
     # sphinx extension
     "python/chemiscope/sphinx/static/chemiscope.min.js",
     # streamlit component
@@ -81,7 +78,7 @@ def needs_npm_build():
 
 
 def run_npm_build():
-    """Build the JavaScript code required by the jupyter widget using npm"""
+    """Build the JavaScript code required by the notebook widget using npm"""
     root = os.path.dirname(os.path.realpath(__file__))
 
     # This file does not exist when building a wheel (for pip installation)
@@ -89,41 +86,16 @@ def run_npm_build():
     # js files.
     if os.path.exists(os.path.join(root, "package.json")):
         subprocess.run("npm ci", check=True, shell=True)
-        # build the labextension first, since it triggers a rebuild of the main
-        # package code
-        try:
-            import jupyterlab  # noqa
 
-            subprocess.run("npm run build:labextension", check=True, shell=True)
-            with open(
-                os.path.join(root, "python", "jupyter", "labextension", "install.json"),
-                "w",
-            ) as fd:
-                install_json = {
-                    "__comment": "metadata for chemiscope jupyterlab extension",
-                    "packageManager": "python",
-                    "packageName": "chemiscope",
-                    "uninstallInstructions": "Use your Python package manager "
-                    + "(pip, conda, etc.) to uninstall chemiscope",
-                }
-                json.dump(install_json, fd)
-        except ImportError:
-            # skip building jupyterlab if it is not installed on the developer
-            # machine
-            warnings.warn(
-                "skipping the lab extension build since jupyterlab is not installed",
-                stacklevel=1,
-            )
-
-            # still build the main code for the sphinx extension
-            subprocess.run("npm run build", check=True, shell=True)
-
+        # build the main code, used by the sphinx extension
+        subprocess.run("npm run build", check=True, shell=True)
         shutil.copyfile(
             src="dist/chemiscope.min.js",
             dst="python/chemiscope/sphinx/static/chemiscope.min.js",
         )
 
-        subprocess.run("npm run build:nbextension", check=True, shell=True)
+        # build the anywidget ES module for the notebook widget
+        subprocess.run("npm run build:anywidget", check=True, shell=True)
 
         subprocess.run("npm run build:streamlit", check=True, shell=True)
 
@@ -141,54 +113,22 @@ def run_npm_build():
 
 
 if __name__ == "__main__":
-    # we need to run npm build outside of the call to setup to be able to get
-    # the full list of files to install for the labextension. Building the
-    # labextension creates unpredictable file names, and `data_file` wants to
-    # explicitly list all files one by one.
-
     if needs_npm_build():
         print("==== rebuilding the JS code with npm, this might take a while ...")
         run_npm_build()
 
[…]
