### 2026-06-24 78e95bd
Added a Claude skill for chemiscope.
