### 2026-06-25 c42c697
Make skill files harness-agnostic
