+++ docs/_downloads/dc40fee6438aac5d0cbaf0566f61cde4/8-explore-with-metatomic.py
@@ -1,0 +1,329 @@
+"""
+.. _chemiscope-explore-metatomic:
+
+Using metatomic models for dataset exploration
+==============================================
+
+In this example, we demonstrate how to create and use a `metatomic`_ model with
+:py:func:`chemiscope.metatomic_featurizer` to extract features from the model, which are
+then displayed using a chemiscope widget. To use this function, some additional
+dependencies are required. You can install them with the following command:
+
+.. code:: bash
+
+    pip install chemiscope[explore]
+
+.. _metatomic: https://docs.metatensor.org/metatomic/
+"""
+
+# %%
+#
+# Firstly, we import necessary packages and read structures from the dataset.
+
+import ase.io
+
+import chemiscope
+
+
+structures = ase.io.read("data/explore_c-gap-20u.xyz", ":")
+
+# %%
+#
+# Using pre-trained models
+# ------------------------
+#
+# Most commonly, you will have an already existing model in metatensor format that
+# you'll want to use for dataset exploration. In this case, you'll have to create a
+# ``featurizer`` function using :py:func:`chemiscope.metatomic_featurizer`.
+#
+# ``metatomic_featurizer`` takes an existing model as input. It can be either a
+# ``AtomisticModel`` instance or a path to a pre-trained model file (here
+# ``"model.pt"``)
+
+featurizer = chemiscope.metatomic_featurizer(model="model.pt")
+
+# %%
+#
+# From here, you can use :py:func:`chemiscope.explore` to visualize the features
+# computed from the structures. For this, we are passing the structures, the
+# ``featurizer`` function, and — as the model computes per-atom properties — the list of
+# environments.
+
+chemiscope.explore(
+    structures=structures,
+    featurizer=featurizer,
+    environments=chemiscope.all_atomic_environments(structures),
+)
+
+# %%
+#
+# Defining a custom model
+# -----------------------
+#
+# Let's now move on and see how one can define a fully custom model to use through the
+# metatomic interface.
+#
+# Here we will use an atom-centered representation, where each atomic environment is
+# represented with the moments of the positions of the neighbors up to a maximal order.
+#
+# The model computes moments up to a specified maximum order :math:`k_{\text{max}}`,
+# computing a representation :math:`F_i^k`
+#
+# .. math::
+#
+#     F_i^k = \sum_{j} \frac{r_{ij}}{r_c}^k
+#
+# where :math:`r_{ij}` is the distance between atom :math:`i` and its neighbor
+# :math:`j`, :math:`k` is the moment order and :math:`r_c` is the cutoff radius.
+#
+# Having computed these moments, the model will take a PCA of their values to
+# extract the three most relevant dimensions.
+
+from typing import Dict, List, Optional  # noqa: E402
+
+import torch  # noqa: E402
+from metatensor.torch import Labels, TensorBlock, TensorMap  # noqa: E402
+from metatomic.torch import (  # noqa: E402
+    AtomisticModel,
+    ModelCapabilities,
+    ModelMetadata,
+    ModelOutput,
+    NeighborListOptions,
+    System,
+)
+
+
+class FeatureModel(torch.nn.Module):
+    def __init__(self, cutoff: float, max_k: int):
+        super().__init__()
+        self.cutoff = cutoff
+        self.max_k = max_k
+
+        self._neighbors_options = NeighborListOptions(
+            cutoff=cutoff, full_list=True, strict=True
+        )
+
+    def requested_neighbor_lists(self) -> List[NeighborListOptions]:
+        # our model requires a neighbor list, that will be computed and provided to it
+        # automatically.
+        return [self._neighbors_options]
+
+    def forward(
+        self,
[…]
