+++ docs/_downloads/ea6d438f77eb9f18b9a2892f5776aa6b/6-explore.py
@@ -1,0 +1,143 @@
+"""
+.. _explore-example:
+
+Exploring dataset with chemiscope
+=================================
+
+The :py:func:`chemiscope.explore` function provides a streamlined way to visualize
+datasets as the low dimensional maps. This approach provides a quick and interactive
+overview of dataset composition and structure without the need to manually implement and
+configure the representation processes. This is particularly useful when the specific
+choice of hyperparameters does not significantly impact the resulting low-dimensionality
+map.
+
+By passing a list of `ase.Atoms <https://ase-lib.org/ase/ase/atoms.html>`_ objects
+(or similar structures from other libraries) to :py:func:`chemiscope.explore`, it is
+possible to generate a chemiscope widget, providing an immediate and intuitive
+visualization of the dataset.
+
+By default, the method uses the `PETMADFeaturizer <https://arxiv.org/abs/2506.19674>`_,
+which computes representations as the `PET-MAD <https://arxiv.org/abs/2503.14118>`_
+features and maps them a low-dimensional MAD dataset latent space.
+
+For more advanced use cases, :py:func:`chemiscope.explore` allows to provide a custom
+function for representation and dimensionality reduction.
+
+To use this function, some additional dependencies are required. You can install them
+with the following command:
+
+.. code:: bash
+
+    pip install chemiscope[explore]
+
+In this example, we will explore basic and advanced use cases, from simple dataset
+visualization to custom featurization.
+
+First, let's import the necessary packages that will be used throughout the examples.
+"""
+
+# %%
+import ase.io
+
+import chemiscope
+
+
+# %%
+#
+# Basic example
+# +++++++++++++
+#
+# This example shows the basic usage of :py:func:`chemiscope.explore`. First, load a
+# dataset of structures as `ase.Atoms <https://ase-lib.org/ase/ase/atoms.html>`_
+# objects. Here, we use the samples from the `MC3D dataset
+# <https://doi.org/10.24435/materialscloud:rw-t0>`_:
+
+structures = ase.io.read("data/explore_mc3d.xyz", ":")
+
+
+# %%
+#
+# Next, pass the structures to :py:func:`chemiscope.explore` to generate an interactive
+# Chemiscope. In this basic case, we provide the featurizer version to be used:
+
+chemiscope.explore(structures, featurizer="pet-mad-1.0")
+
+# %%
+#
+# We can also save the visualization to send it to collaborators or reopen
+# separately with :py:func:`chemiscope.show_input`:
+chemiscope.explore(
+    structures,
+    featurizer="pet-mad-1.0",
+    write_input="mc3d.chemiscope.json.gz",
+)
+
+
+# %%
+#
+# Besides this, it is possible to specify atom-centered environments and properties.
+# Environments can be manually defined as a list of tuples in the format
+# ``(structure_index, atom_index, cutoff)`` or extracted automatically using
+# :py:func:`chemiscope.all_atomic_environments`. We can also configure visualisation
+# settings, such as axis and color properties.
+
+properties = chemiscope.extract_properties(structures, only=["energy"])
+environments = [(0, 0, 3.5), (1, 0, 3.5), (2, 1, 3.5)]
+settings = chemiscope.quick_settings(
+    x="features[1]", y="features[2]", map_color="energy"
+)
+chemiscope.explore(
+    structures,
+    featurizer="pet-mad-1.0",
+    environments=environments,
+    properties=properties,
+    settings=settings,
+)
+
+# %%
+#
+# Example with custom featurizer
+# ++++++++++++++++++++++++++++++
+#
+# For advanced use cases, you can define a custom featurization function. For example,
+# we can describe structures based on their chemical compositions. The function must
[…]
