### 2026-07-06 066417a
update upet and vesin versions
