+++ pyproject.toml
@@ -75,9 +75,9 @@ docstring-code-format = true
 
 [project.optional-dependencies]
 explore = [
-    "pet_mad>=1.4,<2.0",
+    "upet>=0.2.6,<0.3",
     "metatomic-torch >=0.1.12,<0.2",
-    "vesin >=0.3.5,<0.6",
+    "vesin >=0.5.4,<0.6",
 ]
 streamlit = ["streamlit>=1.25"]
 headless = ["playwright>=1.52"]
