### 2026-07-06 6cbe2c8
drop python 3.10 as upet requires >= 3.11
