+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "chemiscope"
 dynamic = ["version"]
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 dependencies = [
     "numpy",
     "anywidget",
