### 2026-07-07 6a674c5
Raise an error if a model request custom inputs
These are not supported in chemiscope at this point
