### 2026-07-07 d1ccc37
Update docs/skills/settings-reference.md
