### 2026-07-07 08f0550
Update docs/skills/settings-reference.md
