+++ docs/skills/settings-reference.md
@@ -81,7 +81,7 @@ covers) or when you need exact field names, defaults, and accepted values.
 Shared by `map.color.palette` and `structure[].color.palette`:
 
 `inferno`, `magma`, `plasma`, `viridis`, `cividis`, `seismic`, `brg`, `bwr`, `rwg`,
-`twilight` (periodic), `twilight dark` (periodic), `hsv` (periodic), `tab10`, `tab20`,
+`twilight (periodic)`, `twilight dark (periodic)`, `hsv (periodic)`, `tab10`, `tab20`,
 `tab20b` (the `tab*` ones are categorical).
 
 ## Notes & legacy quirks
