### 2026-07-07 2e0a881
Update docs/skills/shapes-reference.md
