### 2026-07-13 355cb01
fix draggable windows from going outside of the screen
