### 2026-08-07 d5f5d3e
apply ruff 0.16 formatting and pin ruff version (#591)
see a failed job in CI:
https://github.com/lab-cosmo/chemiscope/actions/runs/30228253301/job/89862163813?pr=590

it fails because the new version of ruff is not happy with one line in
our README
i also pin `ruff==0.16.*` in the tox.ini so future ruff releases do not
break CI unexpectedly
