+++ tox.ini
@@ -42,7 +42,7 @@ commands =
 description = Lint Python code
 package = skip
 deps =
-    ruff
+    ruff==0.16.*
 
 commands =
     ruff format --diff {[testenv]lint_folders}
@@ -52,7 +52,7 @@ commands =
 description = Abuse tox to do actual formatting on all files
 package = skip
 deps =
-    ruff
+    ruff==0.16.*
 commands =
     ruff format {[testenv]lint_folders}
     ruff check --fix-only {[testenv]lint_folders}
