+++ streaming.ts
@@ -1,0 +1,747 @@
+/**
+ * Streaming dataset loader for the web app.
+ *
+ * Loading a large dataset by reading the whole file into a string and calling
+ * `JSON.parse` can exhaust memory or exceed the engine's max string length. This
+ * module instead reads the file as a byte stream (transparently gunzipping
+ * `.gz` files) and tokenizes the JSON incrementally with a small hand-written
+ * parser, so the full text never needs to live in memory at once.
+ *
+ * The bulk of a dataset is its `structures` array. As each structure object is
+ * parsed, its raw JSON text is written (in batches) to an IndexedDB store, and
+ * only a lightweight `UserStructure` placeholder ({ size, data: index }) is kept
+ * in memory. This keeps the resident set small even for files larger than RAM.
+ * Smaller sections (`properties`, `environments`, metadata) are parsed normally.
+ *
+ * The returned `loadStructure(index)` callback lazily reads a structure back
+ * from IndexedDB and parses it on demand (chemiscope only materializes a
+ * structure when it is actually displayed). `release()` drops the IndexedDB
+ * store; only one streamed dataset is kept alive per page (see `activeRelease`),
+ * and the store is also released on `pagehide` and when a new dataset loads.
+ *
+ * Usage from the app:
+ *   if (shouldUseStreaming(file)) {
+ *       const { dataset, loadStructure } = await loadDatasetStreaming(file, (p) =>
+ *           setLoadingProgress(progressMessage(p))
+ *       );
+ *       await visualizer.load({ loadStructure }, dataset);
+ *   }
+ *
+ * `parseJsonWithNaN` is also exported and reused by the app's non-streaming path
+ * to accept the non-standard bare `NaN` tokens that Python's json module emits.
+ */
+
+import type { Dataset, Environment, Property, Structure, UserStructure } from '../src/dataset';
+
+// how many structures go into a single store record
+const BATCH_SIZE = 2048;
+// how many store writes can run in the background before we wait
+const MAX_PENDING_WRITES = 32;
+// drop the consumed part of the parse buffer past this size
+const COMPACT_THRESHOLD = 1 << 16;
+// progress updates are throttled to keep the displayed numbers readable
+const PROGRESS_THROTTLE_MS = 100;
+
+const WHITESPACE = new Set([' ', '\t', '\n', '\r']);
+
+export interface StreamingDataset {
+    dataset: Dataset;
+    loadStructure: (index: number) => Promise<Structure>;
+    release: () => Promise<void>;
+}
+
+export interface StreamingProgress {
+    bytesRead: number;
+    bytesTotal: number;
+    structures: number;
+    phase: 'structures' | 'flushing' | 'properties' | 'environments' | 'finalizing' | 'preparing';
+}
+
+// files at/above this size (MB) stream; smaller ones load fully in memory
+const STREAMING_THRESHOLD_MB = 75;
+
+/** Whether a file should be loaded through the streaming path. */
+export function shouldUseStreaming(file: File): boolean {
+    const isGzipped = file.name.endsWith('.gz');
+    const sizeMB = file.size / (1024 * 1024);
+    const canStream =
+        typeof indexedDB !== 'undefined' &&
+        (!isGzipped || typeof DecompressionStream !== 'undefined');
+    return sizeMB >= STREAMING_THRESHOLD_MB && canStream;
+}
+
+/** Human-readable loading message for a streaming progress update. */
+export function progressMessage(progress: StreamingProgress): string {
+    const percent =
+        progress.bytesTotal > 0
+            ? ` (${((100 * progress.bytesRead) / progress.bytesTotal).toFixed(0)}%)`
+            : '';
+
+    switch (progress.phase) {
+        case 'structures':
[…]
