+++ docs/_sources/python/cli.rst.txt
@@ -1,0 +1,10 @@
+.. _chemiscope-input-cli:
+
+Command line interface
+======================
+
+.. sphinx_argparse_cli::
+   :module: chemiscope.main
+   :func: _chemiscope_input_parser
+   :prog: chemiscope-input
+   :group_title_prefix:

[…]
