+++ docs/_downloads/f3a896bd1641407e60fa3c96a15014d5/10-visualizing-biomolecules.py
@@ -1,0 +1,233 @@
+"""
+Visualize biomolecules with MDAnalysis or chemfiles
+===================================================
+
+This example shows how to visualize biomolecules in chemiscope with `MDAnalysis
+<https://www.mdanalysis.org/>`_ or `chemfiles <https://chemfiles.org/>`_, and
+how to leverage the selection capabilities of these libraries to show only a
+subset of the atoms.
+
+Biomolecules often contain a large number of atoms, which makes the classical
+ball-and-stick representation of molecules hard to be understand.
+The "cartoon" representation focuses on the main structural elements, such as
+backbone atoms, to highlight the secondary structure, and is often more readable.
+"""
+
+import urllib.request
+
+import chemfiles
+import MDAnalysis as mda
+import numpy as np
+from MDAnalysis.analysis.rms import RMSD
+from MDAnalysis.tests.datafiles import GRO_MEMPROT, XTC_MEMPROT
+
+import chemiscope
+
+
+# %%
+#
+# Retrieving the PDB file from RCSB Protein Data Bank
+# +++++++++++++++++++++++++++++++++++++++++++++++++++
+#
+# The `RCSB Protein Data Bank <https://www.rcsb.org/>`_ (RCSB PDB) is a database of
+# crystal structures of proteins, nucleic acids and small molecules. To start with, we
+# will retrieve a structure from the PDB database. Here we choose "10MH", a complex
+# consisting of a protein, a nucleic acid, small molecules, and crystallographic water.
+
+pdb_id = "10MH"
+urllib.request.urlretrieve(
+    f"https://files.rcsb.org/view/{pdb_id}.pdb", f"./{pdb_id}.pdb"
+)
+
+# %%
+#
+# Using MDAnalysis
+# ++++++++++++++++
+#
+# We use `MDAnalysis <https://www.mdanalysis.org/>`_ to read the PDB file, interpreting
+# also the metadata that describes the structure of the protein.
+
+universe = mda.Universe(f"./{pdb_id}.pdb")
+
+# The `chemiscope` takes a `MDAnalysis.AtomGroup
+# <https://userguide.mdanalysis.org/stable/atomgroup.html>`_ as input. You can toggle
+# the cartoon representation in the hamburger menu in the top-right corner of the
+# widget. When the cartoon representation is off, the representation will automatically
+# fall back to the ball-and-stick representation.
+
+ag = universe.atoms
+chemiscope.show(
+    structures=ag,
+    mode="structure",
+    settings=chemiscope.quick_settings(structure_settings={"cartoon": True}),
+)
+
+# %%
+#
+# Selecting atoms of interest
+# ---------------------------
+#
+# The crystallographic water in the structure is not of interest, so we can use the
+# `select_atoms()` method to only show the complex for a cleaner visualization.
+
+sol = universe.select_atoms("not water")
+chemiscope.show(
+    structures=sol,
+    mode="structure",
+    settings=chemiscope.quick_settings(structure_settings={"cartoon": True}),
+)
+
+# %%
+#
+# Using chemfiles
+# +++++++++++++++
+#
+# Alternatively, we can use `chemfiles <https://chemfiles.org/>`_ to read the PDB file.
+
+trajectory = chemfiles.Trajectory(f"./{pdb_id}.pdb")
+
+# Just like with MDAnalysis, we can pass the chemfiles frame directly to chemiscope.
+chemiscope.show(
+    structures=trajectory,
+    mode="structure",
+    settings=chemiscope.quick_settings(structure_settings={"cartoon": True}),
+)
+
+# %%
+#
+# Selecting atoms of interest
+# ---------------------------
+#
[…]
