+++ docs/_sources/examples/11-headless.rst.txt
@@ -1,0 +1,469 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "examples/11-headless.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_examples_11-headless.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_examples_11-headless.py:
+
+
+Headless mode
+=============
+
+This example demonstrates how to use the headless mode of chemiscope to generate
+screenshots programmatically, without running a Jupyter notebook.
+
+We load a dataset of atomic polarizabilities, visualize it, and then
+use the headless mode to generate a curated matplotlib plot. We will show
+snapshots of the two structures in the dataset (with ellipsoids representing
+polarizability) and detailed views of specific atomic environments (colored by
+the polarizability trace) embedded in the plot.
+
+.. GENERATED FROM PYTHON SOURCE LINES 14-27
+
+.. code-block:: Python
+
+
+    import io
+    import time
+
+    import ase.io
+    import matplotlib.pyplot as plt
+    import numpy as np
+    from matplotlib.offsetbox import AnnotationBbox, OffsetImage
+
+    import chemiscope
+    from chemiscope import headless
+
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 28-32
+
+Load the dataset
+----------------
+
+We use the alpha-mu dataset which contains atomic polarizability tensors.
+
+.. GENERATED FROM PYTHON SOURCE LINES 33-36
+
+.. code-block:: Python
+
+
+    structures = ase.io.read("data/alpha-mu.xyz", ":")
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 37-41
+
+Data Processing
+---------------
+
+We compute the atomic polarizability trace and anisotropy for each atom.
+
+.. GENERATED FROM PYTHON SOURCE LINES 42-91
+
+.. code-block:: Python
+
+
+    atomic_trace = []
+    atomic_anisotropy = []
+
+    for structure in structures:
+        # Construct atomic tensors (Voigt notation: xx, yy, zz, xy, xz, yz)
+        # from the individual arrays in the file
+        structure.arrays["alpha"] = np.array(
+            [
+                [axx, ayy, azz, axy, axz, ayz]
+                for (axx, ayy, azz, axy, axz, ayz) in zip(
+                    structure.arrays["axx"],
+                    structure.arrays["ayy"],
+                    structure.arrays["azz"],
+                    structure.arrays["axy"],
+                    structure.arrays["axz"],
+                    structure.arrays["ayz"],
+                    strict=True,
+                )
+            ]
+        )
+
+        alphas = structure.arrays["alpha"]
+        for i in range(len(structure)):
+            # Voigt to 3x3 matrix
+            val = alphas[i]
+            tensor = np.array(
+                [
+                    [val[0], val[3], val[4]],
+                    [val[3], val[1], val[5]],
+                    [val[4], val[5], val[2]],
+                ]
+            )
+
+            evals = np.linalg.eigvalsh(tensor)
+            trace = np.sum(evals)
+            anisotropy = np.abs(evals[2] - evals[0])
+
+            atomic_trace.append(trace)
+            atomic_anisotropy.append(anisotropy)
+
+    properties = {
+        "Atomic Trace": atomic_trace,
[…]
