### 2026-09-22 6ff2131
Added a checker and an error message when WebGL is disabled (#597)
Otherwise the widget crashes with an obscure error

Also fixes an annoying bug with recent webpack
