### 2026-09-23 ceb279f
fix the symbols overlapping the colorbar
