### 2026-09-23 b24a2e5
fix the symbols overlapping the colorbar
