+++ src/map/map.ts
@@ -246,6 +246,14 @@ export class PropertiesMap {
             window.requestAnimationFrame(() => {
                 this._resizePending = false;
                 Plotly.Plots.resize(this._plot);
+
+                // the room left for the colorbar changes with the plot height
+                const len = this._colorbarLen();
+
+                // this runs on every frame of a resize, only redraw if it changed
+                if (len !== this._plot._fullLayout.coloraxis?.colorbar?.len) {
+                    this._relayout({ 'coloraxis.colorbar.len': len } as unknown as Layout);
+                }
             });
         });
         this._resizeObserver.observe(this._plot);
@@ -976,13 +984,16 @@ export class PropertiesMap {
                     this._options.color.min.value = min;
                     this._setScaleStep([min, max], 'color');
 
+                    const colorbar = this._colorbarSettings();
                     this._relayout({
-                        'coloraxis.colorbar.title.text': transformedAxisLabel(
-                            this._title(this._options.color.property.value),
-                            this._options.color.mode.value,
-                            true
-                        ),
-                        'coloraxis.showscale': true,
+                        'coloraxis.colorbar.title.text': colorbar.title,
+                        'coloraxis.showscale': this._showColorbar(),
+                        'legend.maxheight': this._legendMaxHeight(),
+
+                        // null clears old category labels when switching to numeric colors
+                        'coloraxis.colorbar.tickmode': colorbar.tickmode,
+                        'coloraxis.colorbar.tickvals': colorbar.tickvals ?? null,
+                        'coloraxis.colorbar.ticktext': colorbar.ticktext ?? null,
                     } as unknown as Layout);
 
                     // refill the selection range with the new property's bounds
@@ -1001,9 +1012,11 @@ export class PropertiesMap {
                 this._relayout({
                     'coloraxis.colorbar.title.text': undefined,
                     'coloraxis.showscale': false,
+                    'legend.maxheight': this._legendMaxHeight(),
                 } as unknown as Layout);
             }
 
+            this._restyleLegendColors();
             void this._restyleFull();
         });
 
@@ -1039,6 +1052,8 @@ export class PropertiesMap {
             if (this._options.color.select.mode.value !== 'all') {
                 this._restyle({ 'marker.color': this._colors() } as Data, [0, 1]);
             }
+
+            this._restyleLegendColors();
         };
 
         const canChangeColors = (values: number[], changed: string): boolean => {
@@ -1092,13 +1107,13 @@ export class PropertiesMap {
                 this._options.color.min.value = min;
                 this._setScaleStep([min, max], 'color');
 
+                const colorbar = this._colorbarSettings();
                 this._relayout({
-                    'coloraxis.colorbar.title.text': transformedAxisLabel(
-                        this._title(this._options.color.property.value),
-                        this._options.color.mode.value,
-                        true
-                    ),
-                    'coloraxis.showscale': true,
+                    'coloraxis.colorbar.title.text': colorbar.title,
+                    'coloraxis.showscale': this._showColorbar(),
+                    'coloraxis.colorbar.tickmode': colorbar.tickmode,
+                    'coloraxis.colorbar.tickvals': colorbar.tickvals ?? null,
+                    'coloraxis.colorbar.ticktext': colorbar.ticktext ?? null,
                 } as unknown as Layout);
 
                 void this._restyleFull();
@@ -1144,6 +1159,7 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.colorscale': this._options.colorScale(),
             } as unknown as Layout);
[…]
