### 2026-09-23 3f3ea99
remove mace-torch from the docs requirements
