### 2026-09-24 11df676
add workflow to publish npm package on release tags (#603)
