### 2026-09-24 0fe5db9
fix the symbols overlapping the colorbar (#599)
