### 2026-09-24 4da7d15
update scatter3d to accept additional markers (#602)
