### 2026-09-24 1f8b2a9
fix usage of string property with the colorbar + symbols (#601)
when coloring by a string property, the colorbar is label with numerical
values, which is fixed in this PR. when the same property is used for
the symbols and the colors, the legend markers get the colors instead of
showing colorbar

before
<img width="555" height="462" alt="Screenshot 2026-09-23 at 16 00 52"
src="https://github.com/user-attachments/assets/356c925d-2f33-41f7-a89d-e47962cf184a"
/>


after 
<img width="555" height="462" alt="Screenshot 2026-09-23 at 16 00 34"
src="https://github.com/user-attachments/assets/2e056252-
[…]
