+++ src/map/map.ts
@@ -984,14 +984,16 @@ export class PropertiesMap {
                     this._options.color.min.value = min;
                     this._setScaleStep([min, max], 'color');
 
+                    const colorbar = this._colorbarSettings();
                     this._relayout({
-                        'coloraxis.colorbar.title.text': transformedAxisLabel(
-                            this._title(this._options.color.property.value),
-                            this._options.color.mode.value,
-                            true
-                        ),
-                        'coloraxis.showscale': true,
+                        'coloraxis.colorbar.title.text': colorbar.title,
+                        'coloraxis.showscale': this._showColorbar(),
                         'legend.maxheight': this._legendMaxHeight(),
+
+                        // null clears old category labels when switching to numeric colors
+                        'coloraxis.colorbar.tickmode': colorbar.tickmode,
+                        'coloraxis.colorbar.tickvals': colorbar.tickvals ?? null,
+                        'coloraxis.colorbar.ticktext': colorbar.ticktext ?? null,
                     } as unknown as Layout);
 
                     // refill the selection range with the new property's bounds
@@ -1014,6 +1016,7 @@ export class PropertiesMap {
                 } as unknown as Layout);
             }
 
+            this._restyleLegendColors();
             void this._restyleFull();
         });
 
@@ -1049,6 +1052,8 @@ export class PropertiesMap {
             if (this._options.color.select.mode.value !== 'all') {
                 this._restyle({ 'marker.color': this._colors() } as Data, [0, 1]);
             }
+
+            this._restyleLegendColors();
         };
 
         const canChangeColors = (values: number[], changed: string): boolean => {
@@ -1102,13 +1107,13 @@ export class PropertiesMap {
                 this._options.color.min.value = min;
                 this._setScaleStep([min, max], 'color');
 
+                const colorbar = this._colorbarSettings();
                 this._relayout({
-                    'coloraxis.colorbar.title.text': transformedAxisLabel(
-                        this._title(this._options.color.property.value),
-                        this._options.color.mode.value,
-                        true
-                    ),
-                    'coloraxis.showscale': true,
+                    'coloraxis.colorbar.title.text': colorbar.title,
+                    'coloraxis.showscale': this._showColorbar(),
+                    'coloraxis.colorbar.tickmode': colorbar.tickmode,
+                    'coloraxis.colorbar.tickvals': colorbar.tickvals ?? null,
+                    'coloraxis.colorbar.ticktext': colorbar.ticktext ?? null,
                 } as unknown as Layout);
 
                 void this._restyleFull();
@@ -1154,6 +1159,7 @@ export class PropertiesMap {
             this._relayout({
                 'coloraxis.colorscale': this._options.colorScale(),
             } as unknown as Layout);
+            this._restyleLegendColors();
             void this._restyleFull();
         });
 
@@ -1194,9 +1200,12 @@ export class PropertiesMap {
                 name: this._legendNames(),
                 showlegend: this._showlegend(),
             } as unknown as Data);
+            this._restyleLegendColors();
 
             this._relayout({
                 'coloraxis.colorbar.len': this._colorbarLen(),
+                'coloraxis.showscale': this._showColorbar(),
+                'legend.maxheight': this._legendMaxHeight(),
             } as unknown as Layout);
         });
     }
@@ -1528,6 +1537,8 @@ export class PropertiesMap {
 
         // add empty traces to be able to display the symbols legend
         // one trace for each possible symbol
+        const legendColors = this._legendColors();
+
         for (let s = 0; s < this._data.maxSymbols; s++) {
             const data = {
                 name: legendNames[s],
[…]
