+++ app.ts
@@ -1,0 +1,455 @@
+// load bootstrap: this needs to come first so that CSS files are loaded in the
+// right order
+import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+import { Warnings, getByID } from '../src/utils';
+import { Dataset } from '../src/dataset';
+import { version, DefaultVisualizer, Settings } from '../src/index';
+
+import { inflate } from 'pako';
+
+import {
+    loadDatasetStreaming,
+    parseJsonWithNaN,
+    progressMessage,
+    releaseActiveStream,
+    shouldUseStreaming,
+} from './streaming';
+
+// load CSS for the app
+import './app.css';
+
+interface Configuration {
+    /// optional callback to load the structures on demand.
+    loadStructure?: (index: number, structure: unknown) => Promise<any>;
+}
+
+export class ChemiscopeApp {
+    /// Instance of the chemiscope visualizer
+    private visualizer?: DefaultVisualizer;
+    /// Path/URL of the dataset currently displayed
+    private dataset?: string;
+    /// CSS style sheet to hide the setting in loading panel about on-demand
+    /// loading
+    private hideOnDemandStructures: HTMLStyleElement;
+    public warnings: Warnings = new Warnings();
+
+    /**
+     * Create a new instance of the chemiscope application.
+     */
+    constructor(id: string) {
+        // show the version of chemiscope currently running
+        const versionDisplay = getByID('chemiscope-version');
+        versionDisplay.innerText = `version ${version()}`;
+
+        // when the window is resized, change the size available to the info
+        // widget
+        window.addEventListener('resize', updateInfoWidgetHeight);
+
+        // setup the main HTML
+        const root = getByID(id);
+        root.innerHTML = `<div class="container-fluid">
+            <div class="row gx-2">
+                <div class="col-md-6 d-flex flex-column" style="padding: 0 5px">
+                    <div class="ratio ratio-1x1 ratio-map">
+                        <div id="chemiscope-meta" style="z-index: 10"></div>
+                        <div id="chemiscope-map" style="position: absolute"></div>
+                    </div>
+                </div>
+
+                <div class="col-md-6 d-flex flex-column" style="padding: 0 5px">
+                    <div class="ratio ratio-5x7">
+                        <div>
+                            <!-- height: 0 below is a hack to force safari to
+                            respect height: 100% on the children
+                            https://github.com/philipwalton/flexbugs/issues/197#issuecomment-378908438
+                            -->
+                            <div id="chemiscope-structure" style="height: 0"></div>
+                            <div id="chemiscope-info"></div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>`;
+
+        this._setupLoadSaveMenu();
+
+        this.hideOnDemandStructures = document.createElement('style');
+        document.head.appendChild(this.hideOnDemandStructures);
+        this.hideOnDemandStructures.sheet!.insertRule('.hide-on-demand-structures {display: none}');
+        this.hideOnDemandStructures.sheet!.disabled = false;
+    }
+
+    /**
+     * Load the example input with the given name, and display it
+     */
+    public async loadExample(example: string): Promise<void> {
+        let config: Partial<Configuration> = {
+            loadStructure: undefined,
+        };
+        if (example == 'Azaphenacenes') {
+            // example of dynamic structure loading
+            config.loadStructure = async (_, structure: any) => {
+                const url = `examples/${structure.data}`;
+
+                const response = await fetch(url);
+                if (!response.ok) {
+                    throw new Error(
+                        `Failed to fetch structure file "${url}": ${response.status} ${response.statusText}`
+                    );
+                }
+
+                return await response.json();
[…]
